Begin to move common functionality into flower_benchmark.common, align benchmarks
